refactor: update type imports and restructure color-piece and game-movement types
