+++ src/types/game-board.d.ts
@@ -1,1 +1,3 @@
 export type Board = (number | Piece)[][];
+
+export type coordinates = number[];
